Add test for slow-running callback
This test makes a tick callback run for longer than the update interval
by busy-waiting.  The purpose of having it is that the stopwatch should
not rely on being able to get an interval event for each and every
interval.  Instead, it should use the amount of elapsed time as
measured by the system clock.