Add test where stopwatch is paused and restarted
This test is to verify that the stopwatch's time does NOT increase
while the stopwatch is in a "stopped" state.  The codebase already
passes the test as of this commit, but when it is switched to actually
use the system clock to get the elapsed time, this becomes important
to verify.